evaluation/requirements.txt
torch
torchvision
transformers
datasets
huggingface_hub
pandas
Pillow
evaluate
jiwer